Fixes #11 by adding explicit int type casts.

+++ 2.13inch_e-paper/raspberrypi/python/epd2in13.py
@@ -152,7 +152,7 @@ class EPD:
  #  @brief: convert an image to a buffer
  ##
     def get_frame_buffer(self, image):
-        buf = [0x00] * (self.width * self.height / 8)
+        buf = [0x00] * int(self.width * self.height / 8)
         # Set buffer to value of Python Imaging Library image.
         # Image must be in mode 1.
         image_monocolor = image.convert('1')
@@ -166,7 +166,7 @@ class EPD:
             for x in range(self.width):
                 # Set the bits for the column of pixels at the current position.
                 if pixels[x, y] != 0:
-                    buf[(x + y * self.width) / 8] |= 0x80 >> (x % 8)
+                    buf[int((x + y * self.width) / 8)] |= 0x80 >> (x % 8)
         return buf
 
 ##
